# cuda4dnn(slice, concat): reduce slice and concat to copy, more concat fusions