- reduces slice to copy if possible (like in YOLOv4 Tiny)
- reduces concat to copy if possible
- enables more concat fusions

slice from yolov4 tiny         | master            | this PR
--------- | ------------------ | ----------
slice1  | 46us                | 31us
slice2  | 25us                | 17us
slice3  | 19us                | 9us

I think it's possible to fuse slice just like concat but it seems to be a more complex than concat fusion.

<cut/>

### Pull Request Readiness Checklist

See details at https://github.com/opencv/opencv/wiki/How_to_contribute#making-a-good-pull-request

- [x] I agree to contribute to the project under OpenCV (BSD) License.
- [x] To the best of my knowledge, the proposed patch is not based on a code under GPL or other license that is incompatible with OpenCV
- [x] The PR is proposed to proper branch
- [ ] There is reference to original bug report and related work
- [x] There is accuracy test, performance test and test data in opencv_extra repository, if applicable
      Patch to opencv_extra has the same branch name.
- [ ] The feature is well documented and sample code can be built with the project CMake

```
force_builders_only=Custom
buildworker:Custom=linux-4
build_image:Custom=ubuntu-cuda:18.04
```